executor: improve module structure
